requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.5
bark==0.1.5
bitsandbytes==0.46.1
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.39.10
botocore==1.39.10
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1749229842835/work
catalogue==2.0.10
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1752493286596/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560564262/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
contourpy==1.3.2
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
dateparser==1.1.8
decorator==5.2.1
diskcache==5.6.3
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1741969488311/work
Flask==3.1.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
g2pkk==0.1.2
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1745509358281/work
grpcio==1.73.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hangul-romanize==0.1.0
hf-xet==1.1.5
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
huggingface-hub==0.33.4
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jmespath==1.0.1
joblib==1.5.1
jsonlines==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llama_cpp_python==0.3.14
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
msgpack==1.1.1
multidict==6.6.3
murmurhash==1.0.13
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.61.2
numpy==1.26.4
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
opencv-python==4.11.0.86
packaging==25.0
pandas==1.5.3
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1751481992986/work
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.55.0
pysbd==0.3.4
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
regex==2024.11.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1749498106507/work
rich==14.0.0
s3transfer==0.13.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sounddevice @ file:///home/conda/feedstock_root/build_artifacts/python-sounddevice_1748804678685/work
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
trainer==0.0.36
transformers==4.53.3
triton==3.1.0
TTS==0.22.0
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzlocal==5.3.1
umap-learn==0.5.9.post2
Unidecode==1.4.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
